requirements.txt
certifi==2019.3.9
chardet==3.0.4
idna==2.8
numpy==1.16.4
opencv-python==4.1.0.25
Pillow==6.0.0
Pillow-PIL==0.1.dev0
pytesseract==0.2.6
requests==2.22.0
selenium==3.141.0
urllib3==1.25.3
configparser==3.7.4
